requirements.txt
# Kaggle API para descarga de datasets
kagglehub==1.0.0

# Manipulación y procesamiento de datos
pandas==2.3.3
polars==1.39.3

# Machine Learning
scikit-learn==1.8.0
lightgbm==4.6.0
xgboost==3.2.0

# Base de datos
sqlalchemy==2.0.48
psycopg2-binary==2.9.11

# Utilidades
python-dotenv==1.2.2
joblib==1.5.3

# Logging
loguru==0.7.3

# Notebooks
jupyter==1.1.1

# Testing
pytest==9.0.2
pytest-cov==7.1.0
